Return Minterpolate function from interpolate_satdata, fixes
Adjust function interpolate_satdata to return Minterpolate functions for
lon and time as a function of lat rather than interpolated values. This
is needed to have the same interpolated data for flights and satellites.
Use MSessions as first function argument rather than optional argument.
Wrap the loop for interpolating data with pchip in a try-catch-finally
phrase to make sure MATLAB sessions get closed at the end. Throw an
error on failing to interpolate the data.
FIX: In function loadInventory in loadFlightData.jl, remove obsolete
     opening of an MSession that gets never closed.