Redesign struct Intersection, fixes
In TrackMatcher.jl, update docstrings of structs and constructors for
the construction using the original fields in the correct orders, when
modified constructors exist to avoid confusion if structs are used
outside the TrackMatcher package. In the fields of FlightData, allow to
save single values and not just vectors to be able to save the
FlightData struct with intersection data in the Intersection struct.
In struct SatDB, re-allow ZonedDateTime in addition to DateTime for the
use in the 'created' field and make ZonedDateTime the default, when
construction SatDB with the modified constructor.
Redesign struct intersection with fields lat, lon, tdiff, cirrus and
fields sat and flight holding SatData and FlightData with the respective
values at the intersection.
In match.jl, adjust functions intersection and find_intersections to the
new design of struct Intersection with updated function arguments for
SatDB data and a Symbol, whether CPro or CLay data is used.
FIX: In constructors of CLay and CPro, initialise lat/lon as
     Vector{Float64}, not Vector{Any}.
FIX: In SatDB, close the MATLAB session before instantiating a new
     instance, otherwise CLay and CPro will be instatiated with
     'nothing'.