WIP: Reconstruct FlightData
In struct FlightData in TrackMatcher.jl, move all non-metadata fields to
a DataFrame called data with the same column names order as the previous
fields.
In auxiliary.jl, new functions checkcols, checktype, and checkrange to
assure the correct order, names, and types of the columns in
FlightData.data, when constructed with the unmodified constructor.
Make MetaData.source an AbstractString instead of a String.