WIP: Reconstruct FlightDB
In TrackMatcher.jl, add a modified and an unmodified constructor with
database checks to FlightDB. The unmodified constructor checks that only
FlightData of the correct type is assigned to each field and removes
wrongly assigned data in function checkDBtype in auxiliary.jl. The
modified constructor is based on loadFlightDB, but code is revised,
simplified and allows to load the same database type from different
directories (#1).
Remove loadFlightDB from loadFlightData.jl and stop exporting it.