Variable size of SatDB in Intersection; updated kwargs and types
Updated documentation of source code and docstrings.
Revised types for `Real` numbers. Make accuracy in Intersection/
intersection a Float64, precision in functions in match.jl a Float64,
deltat an Int in find_intersections
Introduce kwarg satspan to function intersection, which is passed on to
function find_intersections to instantiate Intersection with a variable
number of data rows saved in SatDB at the Intersection. Adjust function
extract_timespan in auxiliary.jl to a variable timespan.