Add info, warning, and error messages; fixes
In match.jl, rename function get_satranges to findoverlap and variable
satranges to overlap. Add a warning, when no temporal overlap between
flight and sat data was found.
Hand over the flight metadata to function interpolate_satdata and,
subsequently, to MATLAB and issue a MATLAB warning if a flight couldn't
be interpolated about the flight number and database type. Fix issues
regarding looping over segments of the satellite track between flex
points of the trajectory. Rename variable sat to overlap to be
consistent with the remaining code.
Add flight number/source warnings about interpolation failure in MATLAB
to routine interpolate_flightdata. This closes #10.
Adjust parameters in function calls in function interpolation.
In TrackMatcher.jl, make the print statement about the completion of
loading data to FlightDB and info and add a similar info to SatDB.