Revise Intersection data and instantiation
In TrackMatcher.jl, revise struct Intersection and introduce a new
constructor function based on function intersection in match.jl. Add a
new struct XMetadata with metadata holding the flags, computation time,
and time of creation for the Intersection struct.
Intersection is restructured holding 3 DataFrames for the spatial and
temporal coordinates of all intersections in the dataset (coord), the
original flight and satellite data in the vicinity of the intersection
(tracked), and accuracies achieved during interpolation.
In auxiliary.jl, new function get_trackdata to retrieve the original
track data in the vicinity of the intersection based on the previous
constructor of Intersection.
In match.jl, remove function intersection (now constructor of
Intersection) and adjust function find_intersections to new structure of
Intersection. All data is now added to DataFrames rather than individual
Intersection structs and then saved in a struct. The try-catch-finally
phrase is moved inside the loop over all file data rather than having
the whole loop inside the try-catch-finally phrase. This way, only
faulty data gets omitted in the Intersection data and the loop is not
interrupted after the first incident.