Find intersections analytically between 4 closest data points
In TrackMatcher.jl, try to attach the logger to an IO stream named
logfile (which has to be opened and closed in a script outside
TrackMatcher), otherwise use stdout.
In the modified constructor of Intersection, reprint error message and
flight number to screen, if logging level is at debug.
Revise function find_intersections. Besides the minimum distance between
the flight and the satellite track, find the adjacent coordinate pair
with the second least distance and calculate the coordinates of the
intersection analytically between these four data points.
Revised data and error handling in function findoverlap.
Function interpolate_flightdata now returns the Minterpolate function as
well to recalculate the intersection coordinates and time of overflight
once the intersection solution is found analytically in function
find_intersections.
New function timesec to round Unix DateTimes or DateTimes to seconds in
auxiliary.jl.
Updated Manifest.toml.