Make findFiles and remdup functions with bang
Rename findFiles and remdup to findfiles! and remdup! in better
accordance with the Julia naming conventions of functions and adjust
function calls. Stop explicitly assigning the return values.
In loadFlightData.jl, allow missing values in altitude. Make temporal
memory of altitude and speed vectors of type Union{Missing,Float64}.
In TrackMatcher.jl, remove comments of using Revise. This is efficiently
provided by Atom's juno-plus package so that an import of the package in
the source code is not needed anymore.
Use findfiles! in SatDB rather than CLay and CPro, respectively, and
hand over the arrays with the file strings of only the correct type.
This resolves incorrect progress bars during sat data loading.
Adjust structs CLay and CPr