Rename intersection.coord to intersection.data
Rename coord to data in intersection (or idata internally) as the name
coord is not suiting anymore, when feature flags will be added.
Remove feature cirrus from XMetaData as this is a flight-specific
feature and not an overall feature.
Add functions classification and feature_classification to auxiliary.jl
to allow future conversions of the UInt16-encoded feature classification
to human-readable `Symbol` strings. Add function for future
standardisation of units and conversion of feet from the flight data to
kilomenters.