Get feature classification for intersection data, fixes
Add function atmosphericinfo in lidar.jl to return a human-readable
Symbol of the feature classification flag at an intersection in flight
altitude. Return nothing, if flight altitude and lider levels do not
overlap. Adjust the dataframe column type of data in Intersection and
the temporary dataframe in find_intersections to Union{Missing,Symbol}.
In modified constructor of CPro, add warnings about missing data, when
constructing FCF and EC532 column in data.
Add the function call of atmosphericinfo in find_intersections and add
the return value to the idata DataFrame.
FIX: Add metadata as argument in unmodified constructor of CPro data.
     In get_trackdata, hand over metadata to the unmodified constructor
     of CPro.
FIX: Additional filters for the analytic solution of intersections.
     Consider only solutions in a valid latitude/longitude range.
FIX: To get the index of the the nearest measured data point to an
     intersection in the extracted sat and flight data, don't use the
     middle of the data range, but argmin to determine the index as the
     data might be cut-off at the edges. Adjust function get_trackdata
     accordingly.