Add Vector types when constructing CPro data
In the modified constructor of CPro, when Instantiating the data with
the `data` DataFrame use `[Type[]; data...]`, where type is the data
type and data... is the vector with the collected data vectors from the
satellite files. Type needs to be added to avoid a Vector{Any}.
Use Float32 instead of AbstractFloat for EC532.
In find_intersections, use the already defined shorter variable slon
instead of `st.track(ilat)`.