WIP: SatData
New struct SatData with essential data for finding intersections
containing of a data field with a DataFrame with columns time, lat, lon,
and fileindex and a metadata field holding a dict with file names for
each fileindex, the time span of the data, the load time, time of
creation and optional remarks.
Overwrites previous SatMetadata of struct CPro.