Reconstruct metadata for sat data and function checkcols
Remove CProMetadata and move metadata to XMetadata. Add internal
constructors for XMetaData to instantiate lidarrange as tuple or
NamedTuple. Move kwarg lidarrange from CPro constructor to Intersection
constructor as it is now needed in Intersection to construct the
metadata. Adjust instantiation of CPro in extract_timespan and
get_satdata in auxiliary.jl to new CPro stucture.
Hand over hlevels to atmosphericinfo, which cannot be retrieved from the
CPro metadata anymore and adjust function call in get_satdata.
In find_intersections in match.jl, instead of using CLay or CPro as
Xsat, use the respective data fields.
Revise checkcols to modify the data DataFrame and rename to checkcols!.
Add auxiliary functions definebounds!, checkbounds!, findbyname!,
findbyposition!, findbytype!, and correctDF! in auxiliary.jl.  Remove
obsolete auxiliary functions checktype and checkrange. Make bounds a
tuple of pairs with the column number or name pointing to a tuple of
lower and upper bounds and adjust definitions and function calls of
checkcols! in TrackMatcher.jl.
Give additional information about database sizes when flight or sat data
is loaded.
Revise warnings in regard to attached variables.
Add an unmodified constructor for Intersection. Instead of looping over
all data fields of the FlightDB explicitly, use getfield to retrieve all
fields but the last, which is the metadata and construct the FlightData
vector in a more general manner.