Declare SatData type with kwarg; fix
This closes #26.
Introduce kwarg 'type' to SatData, which can set the satellite data type
to ':CLay' or ':CPro'. If set to any other value (or by default), the
type is determined from the first 50 files in the specified folders.
If no files of 'type' are found an empty SatData struct is created.
Add a new external constructor for empty SatData structs.
FIX: Correct automatic type assigment from 'Clay' to 'CLay' in
     constructor of SatMetadata.