Save additional layer data to CLay
Make altmin a Real to allow conversion to km.
Rewrite try-catch phrase to save SatData in a clearer way.
Remove try-catch phrase from CLay/CPro data as the data was already
tested by SatData.
Add columns layer with layer base and top heights, feature with
classification symbols, and OD with feature optical depth to data in
CLay and modify the constructor to check the additional columns.
Overhault the modified constructor to construct the new DataFrame from a
vector of file names, the MATLAB session, the lidarrange, and minimum
flight altitude considered.
Adjust the constructor call of CLay in get_satdata.