Save feature symbols in CPro; redesign atmosphericinfo
Instead of saving the UInt16 representation of feature classification
flags in CPro, save the human-readable feature symbols and rename the
data column from FCF to feature.
Split atmosphericinfo into to methods for finding the feature at flight
level for CPro and CLay data separately. Adjust the function call in
get_satdata depending on the satellite data type.
Add missing columns to the external constructor of CLay.