Load floats to FLightDB with default precision
Introduce new kwarg Float with default value Float32 that defines the
precision of all floating point numbers. Redesign FlightDB, FlightData,
and FlightMetadata to use Float rather than AbstractFloat.
Type FlightData{T<:AbstractFloat} and FlightMetadata{T<:AbstractFloat}.
FIX: In the default constructor of FlightData, check that climb is a
     Vector{<:AbstractFloat} rather than Vector{Int} since the
     conversion to m/s rather than ft/min.
FIX: Correct the progress monitoring in the construction of the 
     FlightAware archive by moving pm.next to the correct outer loop and
     pm.finish outside the correct outer loop.