Add more CPro data; set coarse in get_lidarcolumn to true
Adjust modified constructor of CPro to read in relative humidity and ice
water content from the CALIPSO satellite files.
Adjust the checks in the unmodified constructor for the additional data.
Add empty columns in the empty constructor of CPro.
Set the coarse flag in get_lidarcolumn to true by default as the
majority of the data is coarse except for feature classification flags.