Add more CPro data; revise get_lidarcolumn, fix
Adjust the modified constructor of CPro to read in particulate
depolarization ratio profile at 532 nm, CAD score and day/night flag
from the CALIPSO satellite files.
Adjust the checks in the unmodified constructor for the additional data.
Add empty columns in the empty constructor of CPro.
This closes #30.
Make coarse a kwarg in get_lidarcolumn. Pass only the element type of
the array to be processed by get_lidarcolumn instead of passing the
whole array and inferring the type from it.
FIX: Correct reading in fine mode data with missing values in
     get_lidarcolumn. Instead of converting a Vector convert a Matrix,
     where the 2 dimensions for the 2 levels within an altitude bin are
     considered.