Use haversine function to calculate distances
Depricate the Geodesy package and use Distances instead to calculate the
distance between coordinate pairs. Store distances as tuples of lat/lon
rather than geo.LatLon structs. Calculate the distance with the
haversine function for a perfect fear and introduce function earthradius
that calculates the Earth's radius as a function of latitude and
considers the rotational flattening.
For the analytic solution of the intersection from a linear
interpolation of the 4 closest points, force to use double precision for
most accurate results. This avoids that intersections at segment bounds
get excluded due to rounding errors.
Remove obsolete function get_floatprecision.