Interpolate time linearly
In find_intersections, use a linear interpolation for time using the
distance and time of the two closest points to the intersection.
Add function interpolate_time to do the linear interpolation of time and
function closest_points to find the indices of the points with the
minimum distance to the intersection.
Remove obsolete function timesec.