Revised data interpolation and saving of satellite data
In remdup!, use eps of the current value to set atol of isapprox instead
of using the default rtol value as this is not fine enough for Float32
values.
In get_satdata, use exception handling to save secondary satellite data.
If saving data fails, save an empty struct to the corresponding tracked
field in Intersection and issue a warning.
Update packages in Manifest.toml.