Add parameter expdist to Intersection
Add new keyword argument "expdist" to the modified constructor of
Intersection, which allows setting a maximum distance threshold of the
intersection cooridinates to the next measured flight track point.
Only intersection data with distances to the next flight track point 
below the threshold distance are saved in Intersection.
The default threshold is 100000m.