Fixes and Updates for root find and time interpolation
Correct counter in flight IDs for flights with multiple intersections.
Therefore, set counter to 1 in find_intersections, hand over to addX!,
and increase counter for non-duplicate intersection findings.
For duplicate intersections, only replace columns 2 onwards and leave
the id column unchanged in the temporary dataframes.
In find_intersections, correct the received intersection order from
function findXcoords to "Xf, Xs" rather than "Xs, Xf".
In the main for-loop of find_intersections, enumerate the flighttracks
and use the counter to use only the current flight segment to
interpolate the aircraft flight time at the intersection with function
interpolate_time.
In function findXcoords, use the Krawczyk method rather than the default
Newton method for the interval arithmetids in the roots function of the
IntervalRootFinding package.