Adjust cloud track data structures, find flex points
Make fields time, lat, lon in CloudTrack DataFrame columns and move the
DataFrame to a new field data in accord with the other data structures.
Add column checks on the DataFrame in the constructor of CloudTrack.
Add fields ID, file, flex, and useLON to CloudMetadata and remove the
temporary remarks field.
Add optional remarks to the constructor of CloudDB to be added to the
DBMetadata.
Assign IDs as strings to cloud tracks as
<Int file number>.<Int matrix row number>.
Find segments between flex points and the prevailing trajectory
direction in cloud tracks using preptrack.
Make preptrack a banged function preptrack!, where the DataFrame is not
returned anymore. Rename flight to trajectory to be more general and
clear in the naming scheme, when used with cloud tracks.
In flightdata.jl, don't save FlightData where no track segments between
flex points were found.