Fix calculation of intersections
Overload abs function with new method for CompoundPeriod to be able to
compare the absolute value of a time difference to the maxtimediff
threshold.
In findXcoords, convert xdata and ydata (lat/lon) to Float64 before
finding roots with IntervalRootFinding. This solves issues, where
IntervalRootFinding is stuck in an infinite loop.