WIP: Restructure SatData; introduce SecondarySet/SatSet
Only store single granules in SatData and combine them in SatSet as
subtype of SecondarySet. Introduce a structure for secondary data similar
to primary data.
Remove SatMetadata from SatData and introduce SeconardyMetadata to
SatSet with data from SatMetadata and filenames and time/area extrema
for each granule.
Rename SetMetadata to PrimaryMetadata.
Add function lonextrema that calculates max/min values of longitude for
the eastern and western hemisphere separately. Apply function to
restructured satellite metadata in SatSet constructor and to
FlightMetadata.