WIP: Performance improvements, fixes
Reorganise satellite data and save every granule in a separate DataFrame
for better performance. Revise find time spans and indices in satellite
data.
Process all data in one go in MATLAB and return a data dictionary with
the different variables.
Save the satellite data type as undef<TYpe> instead to show the type was
not set and which type was chosen.
Use checkcols! instead of checkbounds! in unmodified constructor of
SatData.
Correct type in unmodified constructor of FlightData and rename click
back to checkcols!.