WIP: Adjust calculation of cloud intersections, fix
Add obsindex to add_intersections! method for cloud data and adjust type
of sat from SatData to SatSet. Adjust args in call of get_satdata in
add_intersection to new programme structure.
Introduce kwarg structname to constructor of CloudSet that allows a
different default name "cloud" of the top level struct in the cloud
input mat files. Adjust loadCloudTracks and readMAT to new kwarg.
In findoverlap, return an empty range 0:-1 in addition to an empty
DataFrame on a premature return to provide the correct return args.
Correct type of granules in constructor of SatSet from Vector{SatData}
to Vector{SatData{T}}.
Correct name of checkcols! in unmodified constructor of CloudTrack.