Rename FlightSet fields, fix savedir for CloudSet
Rename fields of FlightSet to volpe, flightaware, and webdata.
Rename ID indicators for the source dataset in the intersection data to
V, FA, WD, and CT for VOLPE, FlightAware, web, and cloud tracking data.
Revise internal function and variable names.
Remove continue statement for empty isat variable after find_overlap
call in XData constructor, which is covered by the check of an empty
overlap variable.
Move construction of id in find_intersections after the check of the
time delay to avoid unnecessary computations.
Additionally to ift, check that ist is not 0 to compute spatial and
temporal distances to the next measured data point from an intersection.
Add these checks to the add_intersections method for cloud tracks.
This commit closes #34 and #37.